+++ src/main/java/org/teamseven/tetris/ui/CustomButton.java
@@ -1,4 +1,4 @@
-package org.teamseven.tetris.ui.startmenu;
+package org.teamseven.tetris.ui;
 
 import javax.swing.*;
 import java.awt.*;
