+++ src/main/java/org/teamseven/tetris/ui/IDesign.java
@@ -1,4 +1,4 @@
-package org.teamseven.tetris.ui.startmenu;
+package org.teamseven.tetris.ui;
 
 public interface IDesign {
     void setComp();
